+++ src/utils.ts
@@ -2,7 +2,6 @@ import type { BunFile } from 'bun'
 import {
 	Kind,
 	TAnySchema,
-	TImport,
 	TModule,
 	TransformKind,
 	type TSchema