+++ test/ws/message.test.ts
@@ -258,4 +258,33 @@ describe('WebSocket message', () => {
 		await wsClosed(ws)
 		app.stop()
 	})
+
+	it('handle error', async () => {
+		const app = new Elysia()
+			.ws('/ws', {
+				error() {
+					return 'caught'
+				},
+				message(ws, message) {
+					throw new Error('A')
+				}
+			})
+			.listen(0)
+
+		const ws = newWebsocket(app.server!)
+
+		await wsOpen(ws)
+
+		const message = wsMessage(ws)
+
+		ws.send('Hello!')
+
+		const { type, data } = await message
+
+		expect(type).toBe('message')
+		expect(data).toBe('caught')
+
+		await wsClosed(ws)
+		app.stop()
+	})
 })