+++ example/a.ts
@@ -1,17 +1,24 @@
 import { Elysia, t } from '../src'
+import { req } from '../test/utils'
 
-const app = new Elysia()
-	.model({
-		myModel: t.Object({ num: t.Number() })
-	})
-	.get(
-		'/',
-		({ query: { num } }) => ({ num, type: typeof num }),
-		{
-			query: 'myModel'
+const app = new Elysia().get(
+	'*',
+	async ({ query, request }) => {
+		return {
+			query,
+			url: new URL(request.url).searchParams
 		}
-	)
-	.listen(3000)
+	},
+	{
+		query: t.Object({
+			test: t.Union([t.Array(t.String()), t.String()])
+		})
+	}
+)
+
+app.handle(req('/?test=Test1%20%26%20Test2'))
+	.then((x) => x.json())
+	.then(console.log)
 
 console.log(
 	`🦊 Elysia is running at ${app.server?.hostname}:${app.server?.port}`